IDE:
removed console from remote launch in debug

added options to a vdm project enabling any plugin to store overture settings in a .overture file. This is a request from industry (DESTECS) to be able to keep latex settings together with the project in e.g. SVN

optimized the reconciler a bit to shorten the reconcile loop by pre resolving which parser to load etc.

added synchronized to static variables in SourceUnit reconcile

git-svn-id: https://overture.svn.sourceforge.net/svnroot/overture/trunk@2551 2b2d4d09-c93c-414a-ad04-6e973f5c1cff