When no debug config exists Overture helps the user choosing an entry point using a dialog. Take for example the model:
class A

	operations

	public op1 : () ==> ()
	op1 () == skip;

	public op2 : () ==> ()
	op2 () == skip;

end A

Here Overture suggests op1 and op2 as entry points and if op2 is chosen and a launch configuration is created for this operation. This commit makes this feature work again.

In addition this commit also loosens the requirement for VDM-RT that the entry point must be a Run operation in a World class.