AST Definition Change -------------------------------
type.function and type.operation explicitly for explicit/implicit function/operation makes AstCreator add a case in getType method of the corresponding AST-nodes. Considering an Explicit-function definition as a PDefinition we can set any type for it, but cannot retrieve it again as the getType function dies with a class-cast-exception because of the explicit reference to AFunctionType/AOperationType.

This behaviour is very unexpected and causes the COMPASS type checker to fail as it uses an AErrorType node to mark and track errors…