fixed problem with missing coverage from CT runner in eclipse
the problem was that the ctruntime process were killed before it wrote the covtbl files. Now it waits and signals the IDE when done.

A time out has also been added to the IDE ctruntime socket listner such that the job stops if no connection is made (it uses the debug time out from the debug preference page)