fixed bug: #185 Update of outline causes editor selection to jump while writing
problem was the use of the location searcher (which did not consider the file names) and the fact that the selection in the editor was change even when the current curser position was within the new selection.

Question is now if this also work for windows. The bug no longer occur on Mac with Eclipse Kepler