fixed problem with LexNameToken.equals in the depth first adapters.
The problem is that LexNameToken only can be used in any utils collections within the type chekcer because of the overridden equals implementation.

The depth first adapter uses a Set to avoid infinit recursion do to tree loops from graph fields. This failed because of the equals implementation. The protection is now removed for external node fields, which do not have a depth first implementation anyway. If one is added by a programer then they have to handle the  infinit recursive loop case themself