Change the profiles in ide/pom.xml to
* Autodetect mac and linux as well as windows
* Build both 32/64bit versions for win and linux (only 64bit for mac)

Maven uses OR on activation conditions (despite being documented as AND)
so the original (windows && 64bit) condition was activating on all 64bit
machines.  We now activate only on os type, but build both versions.

Building for all platforms now requires -Pall-platforms -P!mac (or
whichever platform) to disable to autodetected profile (note that the !
needs to be escaped in a bash shell, as -P\!mac).

Fixes #259