modified converagewrite to re-init the properties such that tab stops are restored.
This closes #108 by letting the user set the tab stops in the launc configuration under debugger->parser_tabstop the default value is 1 as used in eclipse. It may be better to stop using tab/detab and just staying with tabs. Eclipse, Word and latex should be able to specify the size of a tab stop.