Improved error messages issued by the Java code generator
If the Java code generator cannot save a Java source files
to disk it will suggest the following solution to the problem
(Windows users):

Operating System: Windows.
If you installed Overture in a location such as "C:\Program Files\Overture"
you may need to give Overture permissions to write to the file system. You can try
run Overture as administrator and see if this solves the problem.