Added early work on a JML generator
The JML generator is implemented  as an extension to the Java vode generator. By intercepting the code generation process the JML generator analyses, modifies and annotates the IR with meta data (JML annotations). Eventually when the annotated IR reaches the Java code generator JML annotated Java code is being produced. The current version of the JML generator works only for VDM-SL and only considers pre and post conditions.