Improving the latex file and coverage generation avoiding multiple calls to the previous LexLocation::removeDuplicates.
* Performance issue:_
Each time the latex file is generated LexLocations may be populated with redundant coverage
information which forces a call to removeDuplicates.

To avoid calling removeDuplicates multiple times the process is encapsulated by creating a snapshot
of the state locations by instantiations of the CoverageUtil class introduced in this commit.

Such class contains methods moved from the LexLocations class, and the removeDuplicates was reshaped.

Cleans createTexFile function.

* Setting dialect only once per project.

In a VDM project the dialect is unique and uniform. So either we have VDM-PP/RT classes
or we have VDM-SL modules. No need to compute the dialect per file...

Moreover the dialect is already passed as a parameter to the makeLatex method... No need
to recompute the value...

Adding TODOs: on functions that need to be understood and documented...